+++ cli/cli/kurtosis_gateway/port_utils/free_port.go
@@ -2,7 +2,7 @@ package port_utils
 
 import (
 	"fmt"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
+	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/stacktrace"
 	"net"
 )